Moved EnrichTutkinnonOsa.ts to a new folder
