+++ shared/models/AiemminHankittuAmmatillinenTutkinnonOsa.ts
@@ -5,7 +5,7 @@ import { EnrichKoodiUri } from "models/EnrichKoodiUri"
 import { EPerusteetVastaus } from "models/EPerusteetVastaus"
 import { LocaleRoot } from "models/helpers/LocaleRoot"
 import { getOsaamispisteet } from "models/helpers/getOsaamispisteet"
-import { EnrichTutkinnonOsa } from "models/EnrichTutkinnonOsa"
+import { EnrichTutkinnonOsa } from "models/Enrichment/EnrichTutkinnonOsa"
 import { TutkinnonOsaViite } from "models/TutkinnonOsaViite"
 import { KoodistoVastaus } from "models/KoodistoVastaus"
 import { AiemminHankitutTutkinnonOsatViews } from "./helpers/AiemminHankitutTutkinnonOsatViews"
