+++ shared/models/HankittavaAmmatillinenTutkinnonOsa.ts
@@ -6,7 +6,7 @@ import { EnrichKoodiUri } from "models/EnrichKoodiUri"
 import { EPerusteetVastaus } from "models/EPerusteetVastaus"
 import { LocaleRoot } from "models/helpers/LocaleRoot"
 import { TutkinnonOsaViite } from "models/TutkinnonOsaViite"
-import { EnrichTutkinnonOsa } from "models/EnrichTutkinnonOsa"
+import { EnrichTutkinnonOsa } from "models/Enrichment/EnrichTutkinnonOsa"
 import { EnrichOrganisaatioOid } from "./EnrichOrganisaatioOid"
 import { Organisaatio } from "./Organisaatio"
 import { TutkinnonOsaType } from "./helpers/ShareTypes"
